backend/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-cors-headers==4.1.0
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
sqlparse==0.4.4
